Refactored simple_service to a handlerTemplate and made it's features available in whole RB
