+++ lib/handlerTemplate.js
@@ -9,38 +9,45 @@ var URI = require('swagger-router').URI;
 var rbUtil = require('../lib/rbUtil');
 var Template = require('../lib/reqTemplate');
 
-function SimpleService(options) {
-    options = options || {};
-    this.spec = {
-        paths: options.paths
-    };
-
-    this.exports = this.processSpec(this.spec);
-}
-
-/**
- * Processes an on_setup property and creates request templates
- * for each resource
- *
- * @param {object} conf an endpoint configuration object
- * @returns {Array} array of request remplates that must be run on service startup
- */
-function processResources(conf) {
-    var resources = [];
-    if (conf.on_setup) {
-        if (Array.isArray(conf.on_setup)) {
-            conf.on_setup.forEach(function(resourceSpec) {
-                Object.keys(resourceSpec).forEach(function(requestName) {
-                    var requestSpec = resourceSpec[requestName];
-                    requestSpec.method = requestSpec.method || 'put';
-                    resources.push(requestSpec);
-                });
-            });
-        } else {
-            throw new Error('Invalid config. on_setup must be an array');
-        }
+function generatePromiseChain(restbase, context, requestChain) {
+    var promise;
+    var currentRequestSpec = requestChain[0];
+    if (currentRequestSpec.length > 1) {
+        promise = P.all(currentRequestSpec.map(function(spec) {
+            return spec.request(restbase, context);
+        }));
+    } else if (currentRequestSpec[0].request) {
+        promise = currentRequestSpec[0].request(restbase, context);
     }
-    return resources;
+
+    if (currentRequestSpec[0].return) {
+        promise = promise || P.resolve();
+        promise = promise.then(currentRequestSpec[0].return(context));
+    }
+
+    if (currentRequestSpec[0].catch) {
+        promise = promise.catch(function(err) {
+            if (!currentRequestSpec[0].catch(err)) {
+                throw err;
+            }
+        });
+    }
+    promise = promise || P.resolve();
+    if (requestChain.length === 1) {
+        return promise;
+    } else if (currentRequestSpec[0].return_if) {
+        return promise.then(function(res) {
+            if (res && currentRequestSpec[0].return_if(res)) {
+                return res;
+            } else {
+                return generatePromiseChain(restbase, context, requestChain.slice(1));
+            }
+        });
+    } else {
+        return promise.then(function() {
+            return generatePromiseChain(restbase, context, requestChain.slice(1));
+        });
+    }
 }
 
 /**
@@ -56,7 +63,7 @@ function compileCatchFunction(catchDefinition) {
                 return '(res["' + catchCond + '"] === ' + option + ')';
             } else if (/^[0-9x]$/.test(option.toString())) {
                 return '(/^' + option.replace('x', '.') +
-                    '$/.test(res["' + catchCond + '"].toString())';
+                '$/.test(res["' + catchCond + '"].toString())';
             } else {
                 throw new Error('Invalid condition ' + option);
             }
@@ -77,7 +84,7 @@ function compileCatchFunction(catchDefinition) {
             condition.push(createCondition(catchCond, catchDefinition[catchCond]));
         }
     });
-    code =  'return (' + condition.join(' && ') + ');';
+    code = 'return (' + condition.join(' && ') + ');';
     /* jslint evil: true */
     return new Function('res', code);
 }
@@ -103,7 +110,7 @@ function validateStep(stepConf) {
     // Returning steps can't be parallel
     if (isReturnStep(stepConf) && Object.keys(stepConf).length > 1) {
         throw new Error('Invalid spec. ' +
-            'Returning requests cannot be parallel. Spec: ' + JSON.stringify(stepConf));
+        'Returning requests cannot be parallel. Spec: ' + JSON.stringify(stepConf));
     }
 
     // Either 'request' or 'return' must be specified
@@ -111,7 +118,7 @@ function validateStep(stepConf) {
         return stepConf[requestName].request || stepConf[requestName].return;
     })) {
         throw new Error('Invalid spec. ' +
-            'Either request or return must be specified. Step: ' + JSON.stringify(stepConf));
+        'Either request or return must be specified. Step: ' + JSON.stringify(stepConf));
     }
 
     // Only supply 'return_if' when 'request' is specified
@@ -119,7 +126,7 @@ function validateStep(stepConf) {
         return stepConf[requestName].return_if && !stepConf[requestName].request;
     })) {
         throw new Error('Invalid spec. ' +
-            'return_if should have a matching request. Step: ' + JSON.stringify(stepConf));
+        'return_if should have a matching request. Step: ' + JSON.stringify(stepConf));
     }
 
     // Only supply 'catch' when 'request' is specified
@@ -127,7 +134,7 @@ function validateStep(stepConf) {
         return stepConf[requestName].catch && !stepConf[requestName].request;
     })) {
         throw new Error('Invalid spec. ' +
-            'catch should have a matching request. Step: ' + JSON.stringify(stepConf));
+        'catch should have a matching request. Step: ' + JSON.stringify(stepConf));
     }
 }
 
@@ -136,8 +143,7 @@ function validateStep(stepConf) {
  * throws Error if some of the rules are not met.
  *
  * Current rules:
- *  - spec must have at least one path
- *  - each request handler spec must have 'on_request' property
+ *  - request handler spec must be an array
  *  - returning steps can't be parallel
  *  - either 'request' or 'return' must be specified in each step
  *  - 'return_if' is allowed only if 'request' is specified in a step
@@ -147,25 +153,15 @@ function validateStep(stepConf) {
  * @param {object} spec service spec object
  */
 function validateSpec(spec) {
-    if (!spec.paths || Object.keys(spec.paths).length === 0) {
-        throw new Error('Invalid spec. Paths are required. Spec: ' + JSON.stringify(spec));
+    if (!Array.isArray(spec)) {
+        throw new Error('Invalid spec. on_request part is required. Spec: ' + JSON.stringify(spec));
     }
-    Object.keys(spec.paths).forEach(function(path) {
-        var conf = spec.paths[path];
-        Object.keys(conf).forEach(function(method) {
-            var handlerConf = conf[method];
-            if (!handlerConf.on_request || !Array.isArray(handlerConf.on_request)) {
-                throw new Error('Invalid spec. ' +
-                    'on_request part is required. Spec: ' + JSON.stringify(conf));
-            }
-            handlerConf.on_request.forEach(validateStep);
+    spec.forEach(validateStep);
 
-            // Last step must not be parallel
-            if (Object.keys(handlerConf.on_request[handlerConf.on_request.length - 1]).length > 1) {
-                throw new Error('Invalid spec. The last step in chain must not be parallel.');
-            }
-        });
-    });
+    // Last step must not be parallel
+    if (Object.keys(spec[spec.length - 1]).length > 1) {
+        throw new Error('Invalid spec. The last step in chain must not be parallel.');
+    }
 }
 
 /**
@@ -255,89 +251,13 @@ function prepareRequestChain(conf) {
     });
 }
 
-/**
- * Processes the service spec and prepares the service to work.
- *
- * @param spec spec object parsed from yaml config
- *
- * @returns {{spec: Object, operations: Object, resources: Array}}
- *          spec: modified spec object with generated operation ids
- *          operations: map with request handlers
- *          resources: array of request templates that should be run on service startup
- */
-SimpleService.prototype.processSpec = function(spec) {
-    var operations = [];
-    var resources = [];
-
-    validateSpec(spec);
-
-    Object.keys(spec.paths).forEach(function(path) {
-        var pathObj = spec.paths[path];
-        Object.keys(pathObj).forEach(function(method) {
-            var conf = pathObj[method];
-            var requestChain = prepareRequestChain(conf.on_request);
-            conf.operationId = method + '_' + path;
-
-            resources = resources.concat(processResources(conf));
-
-            function generatePromiseChain(restbase, context, requestChain) {
-                var promise;
-                var currentRequestSpec = requestChain[0];
-                if (currentRequestSpec.length > 1) {
-                    promise = P.all(currentRequestSpec.map(function(spec) {
-                        return spec.request(restbase, context);
-                    }));
-                } else if (currentRequestSpec[0].request) {
-                    promise = currentRequestSpec[0].request(restbase, context);
-                }
-
-                if (currentRequestSpec[0].return) {
-                    promise = promise || P.resolve();
-                    promise = promise.then(currentRequestSpec[0].return(context));
-                }
-
-                if (currentRequestSpec[0].catch) {
-                    promise = promise.catch(function(err) {
-                        if (!currentRequestSpec[0].catch(err)) {
-                            throw err;
-                        }
-                    });
-                }
-                promise = promise || P.resolve();
-                if (requestChain.length === 1) {
-                    return promise;
-                } else if (currentRequestSpec[0].return_if) {
-                    return promise.then(function(res) {
-                        if (res && currentRequestSpec[0].return_if(res)) {
-                            return res;
-                        } else {
-                            return generatePromiseChain(restbase, context, requestChain.slice(1));
-                        }
-                    });
-                } else {
-                    return promise.then(function() {
-                        return generatePromiseChain(restbase, context, requestChain.slice(1));
-                    });
-                }
-            }
-
-            operations[conf.operationId] = function(restbase, req) {
-                var context = {
-                    request: req
-                };
-                return generatePromiseChain(restbase, context, requestChain);
-            };
-        });
-    });
-
-    return {
-        spec: spec,
-        operations: operations,
-        resources: resources
-    };
-};
-
-
 module.exports = function(options) {
-    return new SimpleService(options).exports;
+    var spec = options || {};
+    validateSpec(spec);
+    var requestChain = prepareRequestChain(spec);
+    return function(restbase, req) {
+        return generatePromiseChain(restbase, {
+            request: req
+        }, requestChain);
+    };
 };
