+++ lib/router.js
@@ -5,7 +5,7 @@ var yaml = require('js-yaml');
 var fs = P.promisifyAll(require('fs'));
 var Template = require('./reqTemplate');
 var rbUtil = require('./rbUtil');
-var BackendRequestHandler = require('./genericReqHandler');
+var buildHandler = require('./handlerTemplate');
 
 var swaggerRouter = require('swagger-router');
 var Node = swaggerRouter.Node;
@@ -126,6 +126,27 @@ function makeRequestTemplate(spec) {
 }
 
 /**
+ * Processes an x-backend-setup config and puts a request templates
+ * for each resource to the node value.
+ *
+ * @param {object} setupConf an endpoint configuration object
+ * @param {object} node a current spec tree node to alter
+ */
+Router.prototype._handleSetupConf = function(setupConf, node) {
+    if (Array.isArray(setupConf)) {
+        setupConf.forEach(function(resourceSpec) {
+            Object.keys(resourceSpec).forEach(function(requestName) {
+                var requestSpec = resourceSpec[requestName];
+                requestSpec.method = requestSpec.method || 'put';
+                node.value.resources.push(requestSpec);
+            });
+        });
+    } else {
+        throw new Error('Invalid config. x-backend-setup must be an array');
+    }
+};
+
+/**
  * Process a Swagger path spec object
  */
 Router.prototype._handleSwaggerPathSpec = function(node, pathspec,
@@ -243,12 +264,14 @@ Router.prototype._handleSwaggerPathSpec = function(node, pathspec,
                     + node.value.path + ':' + methodName);
             }
 
+            var backendSetup = method && method['x-backend-setup'];
+            if (backendSetup) {
+                self._handleSetupConf(backendSetup, node);
+            }
+
             var backendRequest = method && method['x-backend-request'];
             if (backendRequest) {
-                // Set up a templated backend request handler
-                var backendHandler = new BackendRequestHandler(backendRequest);
-                backendHandler.resources.forEach(function(r) { node.value.resources.push(r); });
-                node.value.methods[methodName] = backendHandler.handler.bind(backendHandler);
+                node.value.methods[methodName] = buildHandler(backendRequest);
             } else if (method.operationId) {
                 var handler = operations[method.operationId];
                 if (handler) {
